chore(codex-gpt-5-codex-high): for the cmdk new task, when user selects it, we should sh...
